// src/app/modules/acadimicSemester/acadimicSemester.interface.ts
// export type TMonths =
//   | 'January'
//   | 'February'
//   | 'March'
//   | 'April'
//   | 'May'
//   | 'June'
//   | 'July'
//   | 'August'
//   | 'September'
//   | 'October'
//   | 'November'
//   | 'December';

// export type TAcadimicSemesterName = 'Autumn' | 'Summer' | 'Fall';
// export type TAcadimicSemesterCode = '01' | '02' | '03';

// export type TAcadimicSemester = {
//   name: TAcadimicSemesterName;
//   code: TAcadimicSemesterCode;
//   year: string;
//   startMonth: TMonths;
//   endMonth: TMonths;
// };

//---fourt modul end ---

//-- five modul start ---

export type TMonths =
  | 'January'
  | 'February'
  | 'March'
  | 'April'
  | 'May'
  | 'June'
  | 'July'
  | 'August'
  | 'September'
  | 'October'
  | 'November'
  | 'December';

export type TAcadimicSemesterName = 'Autumn' | 'Summer' | 'Fall';
export type TAcadimicSemesterCode = '01' | '02' | '03';

export type TAcadimicSemester = {
  name: TAcadimicSemesterName;
  code: TAcadimicSemesterCode;
  year: string;
  startMonth: TMonths;
  endMonth: TMonths;
};